Try to start app mock three times before failing
Summary:
In some cases app_mock.py is not able to start mocked app because port is already taken.
This happens because previous processes are not release the ports.
In this cases the app mock will retry max three times and wait a second in between two attempts.

Test Plan: sbt integration:test

Reviewers: zen-dog, jeschkies, timcharper

Subscribers: marathon-dev, marathon-team

Differential Revision: https://phabricator.mesosphere.com/D1034